requirements.txt
aiohttp~=3.8.1
aiohttp-apispec~=2.2.1
aiohttp-cors~=0.7.0
apispec~=3.3.0
async-timeout~=4.0.2
nest_asyncio~=1.5.5
aioredis~=2.0.0
base58~=2.1.0
deepmerge~=0.3.0
ecdsa~=0.16.1
Markdown~=3.1.1
markupsafe==2.0.1
marshmallow==3.5.1
msgpack~=1.0
prompt_toolkit~=2.0.9
pynacl~=1.4.0
requests~=2.25.0
packaging~=20.4
pyld~=2.0.3
pyyaml~=5.4.0
ConfigArgParse~=1.5.3
pyjwt~=2.4.0
pydid~=0.3.6
jsonpath_ng==1.5.2
pytz~=2021.1
python-dateutil~=2.8.1
rlp==0.5.1
unflatten~=0.1
qrcode[pil]~=6.1
cbor2==5.4.2
py-multibase==1.0.3
